fix path error, add terminal checking
